Check column names exist

Internal function to check that supplied data.frames have the correct columns and checks that sample labels are all unique.

Usage

check.cols(ddf.obj, segment.data, observation.data, segment.area)

Arguments

ddf.obj

a ddf object from mrds

segment.data

segment data as defined in dsm

observation.data

observation data as defined in dsm

segment.area

area of segments

Value

nothing, but throws an error if something went wrong

Author

David Lawrence Miller