Density surface modelling

dsm implements spatial models for distance sampling data. Models for detectability can be fitted using packages mrds or Distance. dsm fits generalized additive models to spatially-referenced data. See Miller et al (2013) for an introduction.

Details

Further information on distance sampling methods and example code (including example analyses) is available at https://distancesampling.org/software/Rpackages.html.

For help with distance sampling and this package, there is a Google Group https://groups.google.com/forum/#!forum/distance-sampling.

References

Hedley, S. and S. T. Buckland. 2004. Spatial models for line transect sampling. JABES 9:181-199.

Miller, D. L., Burt, M. L., Rexstad, E. A., Thomas, L. (2013), Spatial models for distance sampling data: recent developments and future directions. Methods in Ecology and Evolution, 4: 1001-1010. doi: 10.1111/2041-210X.12105 (Open Access)

Wood, S.N. 2006. Generalized Additive Models: An Introduction with R. CRC/Chapman & Hall.