Update cookies preferences

Detection function objects when detection is certain

Create a detection function object for strip/plot surveys for use in density surface models.

Usage

dummy_ddf(object, size = 1, width, left = 0, transect = "line")

Arguments

object

numeric vector of object identifiers, relating to the object field in the observation data of the DSM.

size

group size for each observation (default all groups size 1)

width

right truncation

left

left truncation (default 0, no left truncation)

transect

"line" or "point" transect

Author

David L Miller