Convert latitude and longitude to Northings and Eastings

Convert longitude and latitude co-ordinates to kilometres west-east and south-north from axes through (lon0,lat0) using the "spherical law of cosines".

Usage

latlong2km(lon, lat, lon0 = sum(range(lon))/2, lat0 = sum(range(lat))/2)

Arguments

lon

longitude

lat

latitude

lon0

longitude reference point (defaults to mean longitude)

lat0

latitude reference point (defaults to mean latitude)

Value

list with elements km.e and km.n.

Details

WARNING: This is an approximate procedure for converting between latitude/ longitude and Northing/Easting. Consider using projection conversions available in packages sp, sf and rgdal for better results.

Author

Simon N. Wood