Summarize the variance of a density surface model

Gives a brief summary of a fitted dsm variance object.

Usage

# S3 method for class 'dsm.var'
summary(
  object,
  alpha = 0.05,
  boxplot.coef = 1.5,
  bootstrap.subregions = NULL,
  ...
)

Arguments

object

a dsm.var object

alpha

alpha level for confidence intervals (default 0.05 to give a 95\ confidence intervals)

boxplot.coef

the value of coef used to calculate the outliers see boxplot.

bootstrap.subregions

list of vectors of logicals or indices for subregions for which variances need to be calculated (only for bootstraps (see dsm.var.prop for how to use subregions with variance propagation).

...

unused arguments for S3 compatibility

Value

a summary object

See also

Author

David L. Miller